add checks to OWML patcher

+++ UnityMod/OWML/Patcher/Program.cs
@@ -21,24 +21,57 @@ class Program
             }
         }
 
+        if (gameDataPath == "")
+        {
+            System.Console.WriteLine("The provided game path contains no Data folder.");
+            return;
+        }
+
         if (!File.Exists(gamePath + "lincomm.dll"))
         {
-            System.Console.WriteLine("Writing Linux communication library");
-            CopyResourceToPath("lincomm.dll.so", gamePath + "lincomm.dll");
+            
+            if (File.Exists(modPath + "/lincomm.dll.so"))
+            {
+                System.Console.WriteLine("Copying Linux communication library");
+                File.Copy(modPath + "/lincomm.dll.so", gamePath + "lincomm.dll");
+            }
+            else
+            {
+                System.Console.WriteLine("Writing Linux communication library");
+                CopyResourceToPath("lincomm.dll.so", gamePath + "lincomm.dll");
+            }
         }
 
         if (!File.Exists(gameDataPath + "/Plugins/x86_64/OAT_KlakSpout.dll"))
         {
-            System.Console.WriteLine("Writing KlakSpout Plugin");
             Directory.CreateDirectory(gameDataPath + "/Plugins/x86_64");
-            CopyResourceToPath("KlakSpout.dll", gameDataPath + "/Plugins/x86_64/OAT_KlakSpout.dll");
-            File.Copy(modPath + "/KLAKSPOUT-DLL-LICENSE.txt", gameDataPath + "/Plugins/x86_64/KLAKSPOUT-DLL-LICENSE.txt");
+
+            if (File.Exists(modPath + "/KlakSpout.dll"))
+            {
+                System.Console.WriteLine("Copying KlakSpout Plugin");
+                File.Copy(modPath + "/KlakSpout.dll", gameDataPath + "/Plugins/x86_64/OAT_KlakSpout.dll");
+            }
+            else
+            {
+                System.Console.WriteLine("Writing KlakSpout Plugin");
+                CopyResourceToPath("KlakSpout.dll", gameDataPath + "/Plugins/x86_64/OAT_KlakSpout.dll");
+            }
+
+            if (File.Exists(modPath + "/KLAKSPOUT-DLL-LICENSE.txt")){
+                File.Copy(modPath + "/KLAKSPOUT-DLL-LICENSE.txt", gameDataPath + "/Plugins/x86_64/KLAKSPOUT-DLL-LICENSE.txt");
+            }
         }
     }
 
     static void CopyResourceToPath(string resource, string path)
     {
         Assembly asm = Assembly.GetExecutingAssembly();
+
+        if (!asm.GetManifestResourceNames().Contains(resource))
+        {
+            System.Console.WriteLine("No resource "+resource+" exists in assembly.");
+            return;
+        }
 
         using (Stream res = asm.GetManifestResourceStream(resource))
         using (FileStream file = File.Open(path, FileMode.Create))
